+++ src/Pdf++.Core/include/CPPPdf/Text.hpp
@@ -2,5 +2,5 @@
 
 #include <CPPPdf/Text/PdfTextExtractor.hpp>
 #include <CPPPdf/Text/PdfTextSearch.hpp>
-
 #include <CPPPdf/Text/PdfTextDocumentIndex.hpp>
+#include <CPPPdf/Text/PdfTextLayout.hpp>